+++ package.json
@@ -42,6 +42,8 @@
     "babel-preset-react": "^6.23.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "jest": "^23.6.0",
+    "superagent": "^3.8.3",
+    "supertest": "^3.3.0",
     "webpack": "^2.2.1"
   }
 }
